// src/types/CalendarData.ts
// Generated by https://quicktype.io
export interface ICalendarData {
  meetings: IMeeting[];
  deviations: IDeviation[];
}
export interface IDeviation {
  id: string;
  from: string;
  to: string;
  series?: ISeries;
  recurrance?: string;
}
export interface ISeries {
  id: string;
}
export interface IMeeting {
  id: string;
  series?: IDeviation;
  from: string;
  to: string;
  name: string;
  attendees: string[];
}